巴士业务H5初版

+++ src/main/java/com/zhuanche/controller/driverScheduling/DriverSchController.java
@@ -110,7 +110,7 @@ public class DriverSchController {
             if(Check.NuNObj(pageDTO)){
                 return ;
             }
-            Integer total = pageDTO.getTotal();
+            long total = pageDTO.getTotal();
             List<DutyExcelDTO> firstList = new ArrayList<>();
             List<CarDriverDayDutyDTO> result = pageDTO.getResult();
             for (CarDriverDayDutyDTO carDriverDayDutyDTO : result) {
